Refactor the playermodel track and put it in the splinemodel
